requirements.txt
# Python==3.8.19
llama-cpp-python==0.3.0
sentence-transformers==2.2.2
pinecone-client==5.0.1
langchain==0.2.16
Flask==3.0.3
pypdf==5.0.0
python-dotenv==1.0.1
-e .
# Additional dependencies
# aiohttp==3.10.5
# certifi==2024.8.30
# charset-normalizer==3.3.2
# click==8.1.7
# filelock==3.16.1
# huggingface-hub==0.25.1
# idna==3.10
# Jinja2==3.1.4
# MarkupSafe==2.1.5
# numpy==1.24.4
# packaging==23.2
# Pillow==10.4.0
# pydantic==1.10.18
# PyYAML==6.0.2
# requests==2.32.3
# safetensors==0.4.5
# scikit-learn==1.3.2
# scipy==1.10.1
# tokenizers==0.19.1
# torch==2.4.1
# tqdm==4.66.5
# transformers==4.44.2
# typing_extensions==4.10.0
# urllib3==2.2.3
# Werkzeug==3.0.4